Refactors review page part 2
